requirements.txt
browser-use-sdk>=2.0.12
mcp
python-dotenv
dedalus-labs
